Remove wharehouse concept

+++ src/lmjm/model/feed_balance.py
@@ -6,4 +6,4 @@ class FeedBalance:
     pk: str
     sk: str
     measurement_date: str = ""
-    balance_kg: float = 0.0
+    balance_kg: int = 0
