+++ src/lmjm/model/feed_consumption_plan.py
@@ -6,5 +6,5 @@ class FeedConsumptionPlan:
     pk: str
     sk: str
     day_number: int = 0
-    expected_grams_per_animal: float = 0.0
+    expected_grams_per_animal: int = 0
     date: str = ""
